refactor: Extract `handleAddedMessage` into its refresh strategy
